[edit] Made it work, BigDecimal implemented

+++ src/test/java/org/example/TaxParametersTest.java
@@ -14,11 +14,11 @@ public class TaxParametersTest {
 
     @Test
     public void pensionAmountTest() {
-        assertEquals(new BigDecimal("20.00"), pensionAmount(1000));
+        assertEquals(new BigDecimal("20.00"), pensionAmount(BigDecimal.valueOf(1000)));
     }
 
     @Test
     public void unemploymentPaymentTest() {
-        assertEquals(new BigDecimal("16.00"), employeeUnemploymentPaymentAmount(1000));
+        assertEquals(new BigDecimal("16.00"), employeeUnemploymentPaymentAmount(BigDecimal.valueOf(1000)));
     }
 }
